dictionary-check/requirements.txt
requests
beautifulsoup4
lxml
pymorphy3
python-dotenv
playwright